+++ test/upload.php
@@ -1,4 +1,5 @@
 <?php
+header("Access-Control-Allow-Origin: *");
 // http://php.net/manual/fr/features.file-upload.post-method.php
 // Dans les versions de PHP antiéreures à 4.1.0, la variable $HTTP_POST_FILES
 // doit être utilisée à la place de la variable $_FILES.